Howiey medicalnote resolved illnesses + empty differential diagnoses handler (#262)
* Changes resolved illnesses strings

For Patient History > Medical History: If a condition WAS resolved, can say e.g. “Pneumonia (resolved)” or “Pneumonia (resolved 2019)” instead of “Pneumonia resolved” or “Pneumonia resolved (2019)”. In other words put the word 'resolved' inside the parentheses too.

* formatting

prettier formatting :)

* Handling empty differential diagnoses

if there are no comments after the differential diagnosis then there should not be a colon after it. So if the differential diagnosis is "Headache" it should appear as "Headache" rather than "Headache:".

* lint ?

fix lint